refactor(orders): extract algo_builders inline tests to sibling file
Per CLAUDE.md rules 13/14: tests in their own file, modernize touched
modules. Sibling _tests.rs wired via #[path] attribute, contents
unchanged. Prepares the module for upcoming algo builder additions
without mixing test relocation into the feature diffs.
